Update jquery.lazyload.js
Added a new parameter: 'load_all'. When it's set - it will make all the images on the page load at once, whether they are in the viewport or not. Basically it would override the lazyload's idea. It's useful for some things, like when you use a carousel slider on a tablet and you do want the images preloaded.